Use as_ptr_checked in syscalls
Allow system calls ask for operations to be performed on behalf of the
user that roughly correspond to reads and writes over a range.

By checking the supplied syscall arguments could authorise such an
action, we can respect user object boundaries and catch their errors.

On platforms that do nothing in their as_ptr_checked call, this is a
no-op.

Change-Id: I00f96376c0cfa53874966a5f34a566fd2789e89b