kernel/memory_management: Handle process memory configuration validation with MPUs
Until this commit, the kernel checked whether the virtual address space
of a process intersects with its own virtual address space. This
approach does not work for MPUs where the kernel and processes share the
same "virtual" address space. This commit fixes this issue by adding a
config value which tells the kernel whether a MPU is used. The method
used for checking if the process memory configuration is valid has been
changed to take use of the config value.

Signed-off-by: Ioan-Cristian CÎRSTEA <ioan.cirstea@oxidos.io>