boards: tut: dynapp: Add credential checking
Now that we have multiple credential key checking support, this adds
credential checking to the tutorial board.

Because we want to support apps that are not signed as well, this adds a
small wrapper that tells the credential checking machine that its ok to
load apps without a valid credential.