chips/qemu_rv32_virt_chip/uart: ignore spurious interrupts
This is in response to QEMU change 59fad1ebad [1], which fixed their
PLIC implementation to not un-assert interrupts when the source's
level has changed (i.e., implement rising-edge triggered interrupts).

For some (to me still) inexplicable reason, the 16550 UART asserts an
interrupt briefly when Tock starts using it, but that interrupt is not
visible in the peripheral's status / pending interrupt registers. In
this patch, we simply ignore these spurious interrupts.

[1]: https://gitlab.com/qemu-project/qemu/-/commit/59fad1ebadc28995c5d356d097ba6a4d42cae7e7