doc: external dependencies: add understandibility
Refine the qualification around code quality to include aspects like a
narrow focus, common mechanisms, and ability to audit.

The goal with these selection mechanisms is not to create absolute
rules, or to try to ban certain dependencies by prohibiting certain
crates. We are unlikely to define rules or prohibitions we can all agree
on, and we want to avoid creating the incentive for "technically
acceptable" dependencies which maybe we don't want, but don't
technically violate any prohibition.

Instead, these describe the spirit of what we are looking for.