libraries/tock-register-interface: split out into separate repository
This commit removes tock-register-interface from the kernel repository
source tree. From this point onward, this subtree instead lives at
https://github.com/tock/tock-register-interface. Revision b512f97e5eb5
of the tock-register-interface repository corresponds to this remove
tree's contents, modulo added license headers, a CHANGELOG entry, and
a logo for the repository.

This change is motivated (and based on) PR tock/tock#4587, and
implements a long-standing goal of maintaining tock-registers as a
first-class crate with its own stability guarantees, separate from the
kernel. It also follows in line with other kernel dependencies which
we depend on for the Tock kernel, but maintain as part of the Tock
GitHub organization, and thus retain full control over.