capsules/core/alarm: fix `earliest_alarm` to not return expired
This commit fixes the `earliest_alarm` function to not return any
already expired alarms as "earliest". This avoids re-arming the
underlying alarm unnecessarily, if the function is passed one or more
already expired alarms, without an earlier, unexpired alarm.

This bug was found (and fixed) by instructing Gemini 2.5-Pro to
rewrite some of the comments in this file (remaining changes contained
in a subsequent commit). I have carefully read through and further
refined the comments on and in this function to ensure that they
accurately describe the method. I have also reasoned through these
new, changed semantics, to ensure that they don't break the alarm's
system call API and guarantees in any obvious way.

I don't believe that this issue was able to affect userspace in any
way. There are three cases to consider:

- There is at least one unexpired alarm, considered to be firing
  earlier than the time all other expired alarms are expected to fire
  (compared as `ref + dt - now`), relative to `now`.

  Then, `earliest_alarm` works as documented. It will re-arm the
  underlying alarm to fire when the next, unexpired is set to expire.

- There is at least one expired alarm, mistakenly considered to be
  firing earlier than the time all other unexpired alarms are expected
  to fire, relative to `now`.

  In this case, `earliest_alarm` will incorrectly arm the underlying
  alarm for a time _earlier_ than any real, unexpired alarm should
  fire. When the alarm fires, `earliest_alarm` will be called
  again. If that same expired alarm was still in the list of
  Expirations to consider, `now` will have moved past its `ref + dt`,
  and thus this expired alarm will have a high remaining time, causing
  other alarms to be considered instead. Eventually, the function will
  reach the earliest unexpired alarm, and return its expiration to
  correctly set the underlying alarm.

- There are only expired alarms.

  In that case, `earliest_alarm` will re-arm the underlying timer for
  some time in the future. When the alarm fires, `earliest_alarm` will
  re-arm the timer for the "next" expired alarm. If a new, unexpired
  alarm is enqueued through a command, `earliest_alarm` will re-arm
  the alarm according to the above two cases.

In summary, the above should never miss any unexpired alarms,
eventually progressing to them, and will never fire an alarm earlier
than its expiration. I don't think this is a _correctness_ bug, we're
just arming the underlying virtual / hardware alarm spuriously.

Additionally, we eventually get rid of all expired alarms. The
`earliest_alarm` function will still invoke the expired-callback for
all expired alarms, and thus call the `expired_handler` in
`process_rearm_or_callback`. This one will clear the alarm state,
removing it from the list processed on calls to `earliest_alarm`.

Unfortunately, the unit test that was exercising this function was
also incorrect, expecting that an expired alarm be returned as the
earliest alarm.

As an aside, this was the Gemini conversation and prompt that produced
this change: https://g.co/gemini/share/275948bcc867

I specifically requested no code to be changed, and was surprised when
the diff showed a `continue` statement to be added. This makes sense
-- the other comments in that function clearly indicated that the
function _should_ have a continue at this point. However, despite an
arguably pretty impressive feat, it still shows how very dangerous
these tools can be, and the significant impact of such subtle changes,
which are hard to spot in a large diff.