kernel/platform/chip: change `print_state` to take optional `&self`
The changes in tock/tock#4519 mean that a reference to the `Chip`
instance is no longer available in panic contexts invoked from within
a trap handler. However, especially in these cases it is vital to get
some basic information about the chip's state: what execution context
are we in, what are the register contents, ...

If `print_state` requires an `&self` reference, then this function is
not callable in those contexts. Conversely, if it didn't take `&self`,
it would mean that panic messages can only include state that can be
recovered from CPU registers. This, for example, excludes certain PMP
state useful to debug memory-protection related issues.

This PR implements tries to strike a tradeoff: in contexts where the
`Chip` instance is available, it can be used to print useful
software-state, in addition to CPU register state. Otherwise,
`print_state` falls back to just printing CPU register state.

Without this change, trap-handler panics in tock/tock#4519 (such as
because of kernel stack overflows, hard faults, or unhandled
interrupts) will not print any useful CPU state information.