ci-nighly: fix creating GitHub issue
Assinging a team to an issue is not supported. As a result, the
nightly CI is currently failing on every run, but we didn't realize
because the step to create a GitHub issue also fails.

This deliberately does not address the actual failure so we can see
whether the action to create the issue now works.