tock_kernel_layout.ld: place .stack at bottom of SRAM
This is a partial revert of PR tock/tock#4518. That PR incorrectly
swapped the `.stack` and `.relocate` sections in the kernel linker
script, causing the `.stack` section to no longer be placed at the
bottom (the lowest address) in SRAM.

This defeats a core safety mechanism in Tock: when the kernel stack
grows from `_estack` down towards `_sstack` and ultimately overflows
the latter, it should overflow into unmapped / inaccessible
memory. However, with the changes of tock/tock#4518, it would instead
overwrite kernel data.