arch/x86: fix memory initialization routine
Before this change, the assembly routine didn't load the adresses of
the memory ranges to initialize, but instead the _contents_ behind
those addresses.

These happen to be zero-initialized in QEMU, which means that no
`.relocate` data was copied into RAM, and the stack wasn't zeroed. On
platforms that don't zero-initialize their memory, the consequence can
be much more severe and cause arbitrary undefined behavior.

This explains why using `MapCell` inside `SingleThreadValue` did not
work for PR #4696. While the `static DEFCELL` is initialized with a
populated `MapCell`, the `MapCellState` was zero-initialized, which
made Rust believe that this cell is empty.

By switching to `lea`, we load the symbol addresses instead of their
contents, fixing this issue.

[1]: https://github.com/tock/tock/pull/4696